Code style improvements
