+++ QuizApp/Common/Helpers/Primitives+Extensions.swift
@@ -2,19 +2,25 @@ import UIKit
 import Foundation
 
 extension Float {
-    func asCGFloat() -> CGFloat {
+
+    var cgFloat: CGFloat {
         CGFloat(self)
     }
+
 }
 
 extension Int {
-    func asCGFloat() -> CGFloat {
+
+    var cgFloat: CGFloat {
         CGFloat(self)
     }
+
 }
 
 extension Double {
-    func asCGFloat() -> CGFloat {
+
+    var cgFloat: CGFloat {
         CGFloat(self)
     }
+
 }
